Updated delete_false_events.py to only delete events later than a month ago for the selected camera or retain 100 last events, whichever number is greater. Edited the cameras api to include 100 objects in the first page, and the Events api to not filter the events with no detection for any account.

+++ core/management/commands/delete_false_events.py
@@ -1,8 +1,7 @@
 import datetime
-
 from django.core.management.base import BaseCommand
 from django.db.models import Q
-
+from django.utils import timezone
 from core.models import Event
 
 
@@ -16,21 +15,6 @@ class Command(BaseCommand):
         )
 
         parser.add_argument(
-            'date',
-            help='Required, ; all events before this date will be deleted'
-        )
-
-        parser.add_argument(
-            'month',
-            help='Required; month of date'
-        )
-
-        parser.add_argument(
-            'year',
-            help='Required; year of date'
-        )
-
-        parser.add_argument(
             '--delete',
             action='store_true',
             dest='delete',
@@ -40,13 +24,21 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         events = Event.objects.filter(
-                camera__id=options['camera_id'], date__lte=datetime.date(int(options['year']), int(options['month']), int(options['date']))).exclude(Q(species="fire") | Q(species="smoke"))
-        count = events.count()
-        events = events.exclude(status="FEATURED")
+                camera__id=options['camera_id']).exclude(Q(species="fire") | Q(species="smoke")).exclude(status="FEATURED")
+
+        one_month_ago = timezone.now() - datetime.timedelta(days=30)
+        events_in_latest_month = events.filter(date__gte=one_month_ago)
+
+        if events_in_latest_month.count() > 100:
+            events = events.filter(date__lte=one_month_ago)
+        else:
+            events = events[100:]
         message = events.count()
 
-        if options['delete']:
-            events.delete()
-            message = f"Deleted {count}"
 
-        print(f'{message} false events before {options["date"]} - {options["month"]} - {options["year"]} for camera_id {options["camera_id"]}')
+        if options['delete']:
+            if events:
+                events.delete()
+            message = f"Deleted {message}"
+
+        print(f'{message} deletable events for camera_id {options["camera_id"]}')
